hanyang_chatbot/src/requirements.txt
absl-py==0.8.1
appnope==0.1.0
asgiref==3.2.7
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.10.19
botocore==1.13.19
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.9
docutils==0.15.2
en-core-web-sm==2.0.0
entrypoints==0.3
Flask==1.1.1
gast==0.2.2
gensim==3.8.1
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googletrans==2.4.0
grpcio==1.25.0
h5py==2.10.0
hanja==0.13.0
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
JPype1==0.7.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kaggle==1.5.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
konlpy==0.5.1
korean==0.1.8
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.1.1
mecab-python3==0.996.3
mistune==0.8.4
more-itertools==8.0.2
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
Pint==0.10.1
plac==0.9.6
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.7.5
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.6.0
regex==2019.12.9
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.5
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.3.1
srsly==1.0.1
tensorboard==1.10.0
tensorflow==1.10.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==7.0.8
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
ujson==1.35
Unidecode==1.1.1
urllib3==1.24.3
wasabi==0.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.5.0
wrapt==1.11.2
xgboost==0.90
zipp==0.6.0
